Share HTML lexer fixture summary helpers
