+++ code/packages/rust/html-lexer/CHANGELOG.md
@@ -109,6 +109,8 @@ documented in this file.
 - The generic html5lib smoke importer now builds default-wrapper seeded lexer
   cases through the same typed `HtmlLexContext` constructors used by generated
   WHATWG boundary runners.
+- Generated WHATWG lexer fixture runners now share a test-only token summary
+  helper for token, attribute, DOCTYPE, and adjacent-text normalization.
 
 ### Changed
 - Switched the stable `create_html_lexer` and `lex_html` API over from the
