+++ code/packages/rust/html-lexer/tests/whatwg_attribute_edges_test.rs
@@ -1,5 +1,7 @@
-use coding_adventures_html_lexer::{create_html_lexer, Attribute, Token};
+use coding_adventures_html_lexer::create_html_lexer;
 use serde::Deserialize;
+
+mod common;
 
 const WHATWG_ATTRIBUTE_EDGES: &str = include_str!("fixtures/whatwg-attribute-edges.json");
 
@@ -62,11 +64,11 @@ fn whatwg_attribute_edge_cases_match_default_lexer() {
         let actual_tokens = lexer
             .drain_tokens()
             .into_iter()
-            .map(token_summary)
+            .map(common::token_summary)
             .collect::<Vec<_>>();
         assert_eq!(
-            coalesce_adjacent_text_summaries(actual_tokens),
-            coalesce_adjacent_text_summaries(case.tokens.clone()),
+            common::coalesce_adjacent_text_summaries(actual_tokens),
+            common::coalesce_adjacent_text_summaries(case.tokens.clone()),
             "case `{}` token mismatch",
             case.id
         );
@@ -87,80 +89,4 @@ fn whatwg_attribute_edge_cases_match_default_lexer() {
 fn load_suite() -> AttributeEdgeSuite {
     serde_json::from_str(WHATWG_ATTRIBUTE_EDGES)
         .expect("WHATWG attribute-edge fixture should parse")
-}
-
-fn token_summary(token: Token) -> String {
-    match token {
-        Token::Text(data) => format!("Text(data={data})"),
-        Token::StartTag {
-            name,
-            attributes,
-            self_closing,
-        } => format!(
-            "StartTag(name={name}, attributes={}, self_closing={self_closing})",
-            attribute_summary(&attributes)
-        ),
-        Token::EndTag { name } => format!("EndTag(name={name})"),
-        Token::Comment(data) => format!("Comment(data={data})"),
-        Token::Doctype {
-            name,
-            public_identifier,
-            system_identifier,
-            force_quirks,
-        } => doctype_summary(name, public_identifier, system_identifier, force_quirks),
-        Token::Eof => "EOF".to_string(),
-    }
-}
-
-fn doctype_summary(
-    name: Option<String>,
-    public_identifier: Option<String>,
-    system_identifier: Option<String>,
-    force_quirks: bool,
-) -> String {
-    let name = name.unwrap_or_else(|| "null".to_string());
-    match (public_identifier, system_identifier) {
-        (None, None) => format!("Doctype(name={name}, force_quirks={force_quirks})"),
-        (public_identifier, system_identifier) => format!(
-            "Doctype(name={name}, public_identifier={}, system_identifier={}, force_quirks={force_quirks})",
-            public_identifier.unwrap_or_else(|| "null".to_string()),
-            system_identifier.unwrap_or_else(|| "null".to_string())
-        ),
-    }
-}
-
-fn attribute_summary(attributes: &[Attribute]) -> String {
-    if attributes.is_empty() {
-        "[]".to_string()
-    } else {
-        let joined = attributes
-            .iter()
-            .map(|attribute| format!("{}={}", attribute.name, attribute.value))
-            .collect::<Vec<_>>()
-            .join(", ");
-        format!("[{joined}]")
-    }
-}
-
-fn coalesce_adjacent_text_summaries(tokens: Vec<String>) -> Vec<String> {
-    let mut coalesced: Vec<String> = Vec::new();
-    for token in tokens {
-        let Some(text) = token
-            .strip_prefix("Text(data=")
-            .and_then(|text| text.strip_suffix(')'))
-        else {
-            coalesced.push(token);
-            continue;
-        };
-        if let Some(previous) = coalesced.last_mut() {
-            if previous.starts_with("Text(data=") && previous.ends_with(')') {
-                previous.pop();
-                previous.push_str(text);
-                previous.push(')');
-                continue;
-            }
-        }
-        coalesced.push(token);
-    }
-    coalesced
 }
